+++ lib/capybara_accessible_selectors/filters/required.rb
@@ -4,7 +4,7 @@ Capybara::Selector::FilterSet[:capybara_accessible_selectors].instance_eval do
   expression_filter(:required, :boolean, skip_if: nil) do |xpath, value|
     next xpath[XPath.attr(:required) | XPath.attr(:"aria-required") == "true"] if value
 
-    xpath[~XPath.attr(:required) & (~(XPath.attr(:"aria-required") == "true"))]
+    xpath[~XPath.attr(:required) & ~(XPath.attr(:"aria-required") == "true")]
   end
 
   describe(:expression_filters) do |required: nil, **|