+++ lib/capybara_accessible_selectors/selectors/combo_box.rb
@@ -204,7 +204,6 @@ module CapybaraAccessibleSelectors
       find_options[:allow_self] = true if from.nil?
       find_option_options = extract_find_option_options(find_options)
       input = find(:combo_box, from, **find_options)
-      wait_options = { wait: find_options[:wait] }.compact
       if search
         input.set(search, **fill_options)
       else
@@ -216,12 +215,13 @@ module CapybaraAccessibleSelectors
         # Pressing escape will close an open listbox
         input.send_keys(:escape) if has_selector?(:combo_box_list_box, input, wait: 0)
       else
-        listbox = find(:combo_box_list_box, input, **wait_options)
-        option = listbox.find(:list_box_option, with, disabled: false, **find_option_options, &)
-        # Some drivers complain about clicking on a tr
-        option = option.find(:css, "td", match: :first) if option.tag_name == "tr"
-        # Work around occasional Chrome errors
-        option.synchronize(errors: [Selenium::WebDriver::Error::ElementNotInteractableError]) { option.click }
+        input.synchronize(find_options[:wait] || 0) do
+          listbox = find(:combo_box_list_box, input)
+          option = listbox.find(:list_box_option, with, disabled: false, **find_option_options, &)
+          # Some drivers complain about clicking on a tr
+          option = option.find(:css, "td", match: :first) if option.tag_name == "tr"
+          option.click
+        end
       end
       input
     end