+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Remove ruby 3.1 support. Minimum supported Ruby version is now 3.2
+- Extend `Capybara::Node::Simple` with the `role` attribute to support the
+  `:role` filter
 
 ## v0.12.0